zeroshot_code/local_models/requirements.txt
pandas
python-dotenv
lmstudio
